# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "표준 에러 응답. 후속 라우터가 ``responses={409: {\"model\": ErrorResponse}}``로 참조.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GeocodeResult": {
        "description": "주소 검색(지오코딩) 결과 항목(AC2).\n\n카카오 주소검색 ``documents[]`` 한 항목의 매핑 결과다 — ``lat=y``·``lng=x``(카카오는\nx=경도·y=위도), ``admin_dong_code``는 지역 ``b_code``(Dev Notes 결정). 프론트가 후보\n중 하나를 선택해 ``RoomCreateRequest``의 ``lat``/``lng``/``admin_dong_code``로 실어 보낸다.",
        "properties": {
          "address": {
            "title": "Address",
            "type": "string"
          },
          "admin_dong_code": {
            "title": "Admin Dong Code",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "address",
          "lat",
          "lng",
          "admin_dong_code"
        ],
        "title": "GeocodeResult",
        "type": "object"
      },
...
      "RoomCreateRequest": {
        "description": "공간 등록 요청(provider 전용 쓰기, AC1).\n\n범위·enum 위반은 Pydantic 검증 → 1.5 핸들러가 422로 단일화(P3 1차 차단 + 2.1 defer 회수).\n좌표(``lat``/``lng``)·``admin_dong_code``는 클라이언트가 ``GET /rooms/geocode`` 결과에서\n선택해 실어 보낸다(MVP client-trusted — 인증된 provider + DB CHECK 방어, Dev Notes 참조).",
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "admin_dong_code": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Admin Dong Code",
            "type": "string"
          },
          "amenities": {
            "items": {
              "enum": [
                "parking",
                "projector_tv",
                "coffee",
                "whiteboard",
                "wifi",
                "etc"
              ],
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "business_hours": {
            "items": {
              "$ref": "#/components/schemas/BusinessHoursInput"
            },
            "minItems": 1,
            "title": "Business Hours",
            "type": "array"
          },
          "capacity": {
            "minimum": 1.0,
            "title": "Capacity",
            "type": "integer"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "price_per_hour": {
            "minimum": 0.0,
            "title": "Price Per Hour",
            "type": "integer"
          },
          "room_type": {
            "enum": [
              "open",
              "private"
            ],
            "title": "Room Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "price_per_hour",
          "capacity",
          "room_type",
          "lat",
          "lng",
          "admin_dong_code",
          "business_hours"
        ],
        "title": "RoomCreateRequest",
        "type": "object"
      },
...
      "RoomMapItem": {
        "description": "룸 핀 렌더용 최소 메타 — 첫 진입 지도 좌표 공급 (Story 3.2, AC1·AC2).\n\n각 항목은 한 활성 룸의 **핀을 찍는 데 필요한 최소 정보**(좌표·이름)다. 프론트가 이\n목록을 ``GET /rooms/availability``(``RoomAvailability``)와 ``room_id``로 인메모리\n조인해 핀 좌표(이 응답)에 가용성 색(저쪽 응답)을 입힌다.\n\n**이 네 필드만** 둔다 — 가격·영업시간·부대시설·수용·주소·``provider_id``·\n``created_at``은 **싣지 않는다**. 상세는 바텀시트(Story 3.3)의 신선 단일 조회 책임이고,\n공개 엔드포인트라 ``provider_id`` 노출을 피하며, 과조회를 막는다(L155 신선도 정책).\n\n**키 이름 = ``room_id``**(``RoomAvailability.room_id``와 동일) → 두 공개 GET을\n``room_id``로 조인하기 위함이다. ``Room.id``를 ``room_id``로 노출하므로\n``from_attributes`` 자동 매핑에 의존하지 않고 서비스에서 명시 생성한다\n(``RoomMapItem(room_id=r.id, ...)`` — ``RoomAvailability`` 명시 생성 선례).\n\n**와이어 규약:** snake_case 그대로(``room_id``·``name``·``lat``·``lng``). 시각 필드가\n없어 ``field_serializer``가 불필요하다.",
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "room_id": {
            "format": "uuid",
            "title": "Room Id",
            "type": "string"
          }
        },
        "required": [
          "room_id",
          "name",
          "lat",
          "lng"
        ],
        "title": "RoomMapItem",
        "type": "object"
      },
      "RoomPublic": {
        "description": "공간 등록 성공 응답(룸 리소스). ``created_at``은 ``...Z``(1.7 ``UserPublic`` 패턴).",
        "properties": {
          "admin_dong_code": {
            "title": "Admin Dong Code",
            "type": "string"
          },
          "amenities": {
            "items": {
              "type": "string"
            },
            "title": "Amenities",
            "type": "array"
          },
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "title": "Lng",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_per_hour": {
            "title": "Price Per Hour",
            "type": "integer"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "room_type": {
            "title": "Room Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "price_per_hour",
          "capacity",
          "room_type",
          "amenities",
          "lat",
          "lng",
          "admin_dong_code",
          "is_active",
          "created_at"
        ],
        "title": "RoomPublic",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DeskNow 백엔드 API (Story 1.2 골격).",
    "title": "DeskNow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "description": "활성 룸의 핀 메타(``{room_id, name, lat, lng}``)를 반환한다 → 200(공개, AC1·AC2).\n\n첫 진입 지도 핀 좌표 공급(FR-4). **인증 없음**(탐색 첫 화면 핀은 비로그인 — PRD §FR-2,\n``/availability``와 동일 근거). 가용성(핀 색)은 ``/availability``, 룸 상세는 바텀시트\n(Story 3.3)가 책임진다 — 이 응답은 핀 좌표·이름만 싣는다(과조회 방지). 컬렉션 루트\n``GET \"\"``는 ``POST \"\"``(등록·메서드 다름)·``GET /{room_id}``(동적·경로 세그먼트 다름)와\n충돌하지 않으나, 일관성을 위해 정적 그룹 근처·``PATCH /{room_id}`` 앞에 선언한다.\noperationId = ``rooms_list_rooms``(1.9 규약, 도메인 내 유일).",
        "operationId": "rooms_list_rooms",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomMapItem"
                  },
                  "title": "Response Rooms List Rooms",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "description": "공간을 등록한다 → 201 + RoomPublic(provider 전용, AC1·AC5).\n\n인증된 provider의 ``user_id``를 ``provider_id``로 ``create_room`` 호출. 제공자당 1개\n초과는 409, 자정 넘김 영업시간 등 검증 실패는 422(Pydantic→1.5 핸들러).",
        "operationId": "rooms_create_room",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/geocode": {
      "get": {
        "description": "주소를 검색해 좌표·지역 코드 후보를 반환한다 → 200(provider 전용, AC2).\n\n카카오 REST 키는 백엔드 전용(NFR-6)이라 이 프록시를 경유한다. 업스트림 실패는 502.\n결과 0건은 정상(빈 리스트). 정적 경로라 향후 ``/{room_id}``보다 먼저 선언한다.",
        "operationId": "rooms_geocode_address",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeocodeResult"
                  },
                  "title": "Response Rooms Geocode Address",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        },
        "summary": "Geocode Address",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/mine": {
      "get": {
        "description": "provider 본인 룸 상세를 반환한다 → 200(provider 전용 — 등록/수정 폼 prefill).\n\n인증된 provider의 ``user_id``로 본인 룸(0..1)을 조회한다. 없으면 404 ``ROOM_NOT_FOUND``\n(등록 안 함 → 폼 생성 모드). 정적 경로라 동적 ``/{room_id}``보다 **먼저** 선언한다(미선언 시\n\"mine\"이 room_id로 잡혀 422). booker/admin은 403, 미인증 401(``_require_provider``).",
        "operationId": "rooms_get_my_room",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRoomDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Get My Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/search": {
      "get": {
        "description": "선택 지역/반경(또는 전체)의 활성 룸을 신선 슬롯과 함께 **한 페이지** 반환 → 200(공개·AC1·AC4 + F).\n\n목록 행(이름·가격·룸형태·부대시설 + 예약 가능 배지용 신선 ``remaining_slots``). 두 검색방식:\n\n- **지역(3.4):** ``region_code`` 미지정=지역 필터 없음, 지정=시군구/동 레벨 코드로 필터.\n- **반경(3.5):** ``lat``·``lng``가 **둘 다** 있을 때만 반경 필터 적용(부분 좌표=미적용).\n  ``radius_km`` 미지정 시 기본 3km, 결과는 가까운 순 정렬. ``region_code``와 동시=교집합.\n\n**페이징(F):** 거리/지역 계산 정렬이라 **offset 커서**(``CursorPage`` 봉투 — ``items`` +\n``next_cursor``). ``limit``(기본 20·최대 100)·``cursor``는 쿼리, 손상/음수 커서는 422.\n**미매핑/미존재 region_code=빈 페이지 200**(에러 계약 없음 — graceful). 범위 밖 파라미터\n(``lat∉[-90,90]``·``lng∉[-180,180]``·``radius_km≤0`` 또는 50 초과)=**422**(Query 검증→1.5).\n**인증 없음**(탐색은 비로그인 — PRD §FR-2). 정적 경로라 ``/{room_id}``보다 **먼저 선언**한다.\noperationId = ``rooms_search_rooms``(1.9, 도메인 유일).",
        "operationId": "rooms_search_rooms",
        "parameters": [
          {
            "in": "query",
            "name": "region_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region Code"
            }
          },
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "in": "query",
            "name": "lng",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lng"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 3.0,
              "exclusiveMinimum": 0,
              "maximum": 50,
              "title": "Radius Km",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_RoomListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Search Rooms",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/v1/rooms/{room_id}": {
      "get": {
        "description": "단일 룸 신선 요약(바텀시트, Story 3.3) → 200(공개, AC1·AC4).\n\n가격·수용·룸형태·부대시설·영업시간 + ``derive_slots`` 재사용한 **신선 ``remaining_slots``**.\n**인증 없음**(탐색 첫 화면 핀 탭 — PRD §FR-2, ``/availability``·``GET \"\"``와 동일 근거).\n미존재/비활성=404 ``ROOM_NOT_FOUND``. ``provider_id`` 미노출(공개·과조회 방지). 동적 라우트라\n정적 경로 뒤·``PATCH /{room_id}`` 근처 선언. operationId = ``rooms_get_room``(1.9, 도메인 유일).\n4.2 상세 화면이 이 엔드포인트를 후기/예약 UI로 확장한다.",
        "operationId": "rooms_get_room",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "description": "본인 소유 공간을 부분 수정한다 → 200 + RoomPublic(provider 전용, AC1·AC4·AC5).\n\n``model_dump(exclude_unset=True)`` 부분 수정(PATCH 시맨틱) — 미제공 필드는 불변,\n``business_hours`` 제공 시 전체 교체. 소유권은 **서비스에서 최종 강제**(타인/미존재\nroom_id는 404 ``ROOM_NOT_FOUND``로 합침). 자정 넘김·범위 밖 값 등 검증 실패는 422.\noperationId = ``rooms_update_room``(1.9 규약, 도메인 내 유일).",
        "operationId": "rooms_update_room",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Update Room",
        "tags": [
          "rooms"
        ]
      }
    },
...
    "/api/v1/rooms/{room_id}/slots": {
      "get": {
        "description": "활성 룸의 ``date``(ROOM_TZ) 1시간 슬롯을 상태별로 + 다음 빈 날짜를 반환한다 → 200(공개, AC1).\n\n그날 1시간 슬롯을 상태별(``available`` 가용/``past`` 지난 시간/``reserved`` 예약됨)로,\n그리고 그날 다음부터 30일 내 첫 빈 날(``next_available_date``)을 함께 싣는다. ``derive_slots``\n(2.1) 재사용 — ``reserved`` 상태는 **Story 4.9 예약 차감 연결 전까지 미발생**. **인증 없음**\n(상세·슬롯은 비로그인 — PRD §FR-2, ``GET /{room_id}``와 동일 근거). 미존재/비활성=404\n``ROOM_NOT_FOUND``. ``date`` 누락/오형식=422(Query 파싱→1.5 핸들러).\n\n**라우팅 순서:** ``/{room_id}/slots``는 2-세그먼트라 정적 ``/geocode``·``/availability``·\n``/regions``·``/search``·``GET \"\"`` 및 1-세그먼트 동적 ``/{room_id}``와 충돌하지 않는다(경로\n깊이 다름) — 동적 ``/{room_id}`` 그룹 근처에 둔다. operationId = ``rooms_get_room_slots``(1.9,\n도메인 유일) → SDK ``roomsGetRoomSlots``.",
        "operationId": "rooms_get_room_slots",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Room Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomSlotsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Room Slots",
        "tags": [
          "rooms"
        ]
      }
    }
  }
}